fix(solana): fix broken smoke test
The smoke test was expecting the mcm.so file due to the config in `blockchain_a.solana_programs` , since we dont provide any program to deploy in the smoke test, we should not define that config.

I have also updated the default port for solana to be 8999 as that is the default port for solana test validator and it is what users would expect.